Allow for Deterministic, Reproducible Builds with sorted inputs
The annotation processor uses a RequestCodeProvider to assign unique
request codes to each annotated @NeedsPermission method. This shared
state collides with the non-deterministic nature of the processor's
RoundEnvironment, since we can't guarantee the processing order by default.

Here, we sort the affected sets of data before starting the iterations.
First, the annotated @RuntimePermissions classes need to be in the same
order every time we build. Second, inside each annotated class, the
annotated @NeedsPermission methods also need to be in the same order.
Using simple "sortedBy" predicates, this is achieved.

fixes #318