Allow Artifactory publications from both Android & Kotlin-only modules
When applying the Artifactory build script, a Project is now required
to specify an extension property named "publicationName". This is used
to link the Artifactory upload process to a Maven publication, added to
each module individually through the novoda-bintray-release plugin.
Since Android Libraries & "pure JVM" modules are handled differently,
the resulting publications are named differently as well. With this
change, each module can define its own publication to be synched with
Artifactory.