Re-introduce old Lint Detectors based on JavaPsiScanner & modularize
Our IssueRegistry now operates on a List of "Provider" implementations,
which serve as candidates when it comes to providing a list of Issues
to detect via Lint. By default, UastScanner implementations are queried
first, and a fallback to the old JavaPsiScanner classes is made when
UastScanner is absent on the classpath.

The unit tests for the IssueRegistry have been extended to account for
changing Provider implementations with varying availability. Also, issue
creation has been moved into a new class `IssueFactory` to avoid
redundant re-definition of Issues for each Provider.