Replace deprecated "com.google.android" dependency with android.jar
A small custom Gradle plugin is used to look up the user's android.jar
and provide its path as a file dependency to Gradle scripts. With this,
we can finally get rid of the ancient dependency on the Android SDK
present in the processor's test scope, which prevented us from specifying
the Marshmallow Build Constant inside the library. Instead, the plugin
obtains configuration parameters to search & find the user's SDK installation
and grab a reference to the android.jar, which contains the more up-to-date
classes for the processor.